Branching testing; Temporary settings modification
Added a simple test for the branching settings.
This test just vets the ability to use mixed reader levels, ['xs', 'branching'], and storing the full set of variables on the reader object.

Also added the ability for the UserSettingsLoader to be used in a with/as context. Any changes made to the loader during the with/as context will be restored upon exiting the block. This is great for testing, since all the readers share the same settings object, and many tests alter the settings.

Signed-off-by: Andrew Johnson <1drew.e.johnson@gmail.com>