Expand variable groups; XSReader base; Temporary settings  (#27)
* Expanding variables - WIP

Added file variables.yaml to settings directory. This file will serve to faciliate the expansion of select keywords into variables to be scraped from results and branching files.

Unit testing on the function complete. Need to add higher level functionality so that branching and results reader can get the relevant settings.

Signed-off-by: Andrew Johnson <1drew.e.johnson@gmail.com>

* Branching testing; Temporary settings modification

Added a simple test for the branching settings.
This test just vets the ability to use mixed reader levels, ['xs', 'branching'], and storing the full set of variables on the reader object.

Also added the ability for the UserSettingsLoader to be used in a with/as context. Any changes made to the loader during the with/as context will be restored upon exiting the block. This is great for testing, since all the readers share the same settings object, and many tests alter the settings.

Signed-off-by: Andrew Johnson <1drew.e.johnson@gmail.com>