Branching reader now uses homog universes; Testing
Incorporated changes from PR #35 so that the variables are stored in the homogenized universe containers.

Since the variables requested in reader.settings['variables'] may or may not exactly appear in the SERPENT output, a handy method _checkAddVariable was added to XSReader class. This will perform the following checks to see if a variable should be stored: No variables requested -> True; variable directly appears in requested variables -> True; User has requested that B1XS be stored and variable appears without B1_ -> True; User has requested that infinite xs be stored and variable appears without INF_ -> True; Otherwise, return False.

This check is needed because we don't store ALL of the variable names in our settings dictionary, because there are a lot of double counts (INF_TOT and B1_TOT)

Also added testing for the branching reader

Signed-off-by: Andrew Johnson <1drew.e.johnson@gmail.com>