Function to create multiple files using unique seeds
Allows the user to specify number of files to create, starting seed for builtin RNG, output directory, and a rewrite option. If True, the entire input file is copied as a new file into the output directory. If False, the new file uses the include functionality to link the original file to the new file. This results in very short files, only containing the include line and the new seed