List sub-command shows default settings
New subcommand will print the default settings to stdout.
Supports pattern matching to look only for settings that
match some pattern, e.g. start with <dep>.

No verbose/quiet commands given -> description and default shown.
If quiet, only default shown.
If verbose and the setting has options that are not the default,
then the options are printed as well.