Adding multiple v/q flags does not break parser
The previous commits used the direct number of v/q flags to look into the mapping dictionary with the number of flags as the key. If you entered -vvvvvvvv, your would create a KeyError. Now, the level is set to the minimum of the number of flags, and the maximum level in the mapping dictionary