User specified seed length
The user can now specify the size of the random seeds from the
command line, and the writer function will convert that to
the proper number of bits. I did a study to find a linear model
between the number of bits used and the average size of the number.