Abstracted Detector class; Process sampled detector errors
The existing Detector class was refactored to inherit from a
DetectorBase class. This base class contains methods for plotting
and slicing, while the Detector class is utilized by the
DetectorReader and reshaped tally data from files. The
DetectorSampler utilizes a SampledDetector class which inherits from
DetectorBase, and computes average tally data, scores, and
uncertainties.
The uncertainty stored on the SampledDetector object is the
root of the sum of the squared absolute deviations from all files,
divided by the number of files.