SampledContainer base class for SampledDetector
Added a SampledContainer object that serves as a mixin for
the SampledDetector. The SampledContainer is responsible
for verifying that the user is not attempting to load
a number of contains that superceeds the preset amount and
raising warnings if the number of cloned containers does not
equal the number that was specified. Each subclass is required to
implement the methods to obtain average values, and load directly
from an incoming container.