DepletionSampler reads and processes; Setting to free data post-read
DepletionSampler now processes data from all subsequent parsers,
and creates uncertainties for both material data and temporal metadata.
For two files of identical setup with different random seeds, the
ordering of isotope names and zais will be identical, but one or both
of the days/burnup metadata vectors will be different. Since SERPENT
normalizes to either the length of the burnup step in burnup or time
based on the flux at the previous step, and the point of the random
seeds is to have different fluxes, the length of the non-normalized
value could be different.

Added a free method to the Sampler base class that will clear up
all memory associated with the other parsers and their containers.
This first calls a reader-specific free, where the containers are
cleared up. There is a setting that controls this operation,
<samplers.freeAll>, and that setting defaults to False. Subsequent
work today will add the functionality for a spread plot, to plot
the average value and then that value for each individual parser
or container. However, reading in thousands of files could result
in a very large memory requirement, and if the user wants, we clear
that out. The free method is a public method on the Sampler and the
SampledContainer objects, so the user can easily perform this action
at their discretion. An info statement lets the user know if the free
method is automatically performed.