Plot method for sampledDepletedMaterial
SampledDepletedMaterial now plots, with both x and y error,
the same quanties as the standalone DepletedMaterial.

Fixed a minor bug where, if no isotope names were passed onto
the DepletedMaterial plot method, the labeling would fail.
Now, since all isotopes are by default added to the plot if
names is None, the plot labels are the names of all isotopes.