The documentation commit
Updated docstrings for depletion and detector readers and
containers, added docstrings for base containers, and samplers.
Since a lot of the documentation is repeated throughout
readers and containers, the docstrings for almost all of these
objects and their public methods is built by taking, for example,
attributes and parameters from the base classes. This approach
does not appear very well directly in the file, but renders
in a more complete way in interactive sessions or in the manual.