Removed build badge from pdf documentation; Updated pdf
Recently encountered an issue where the sphinx latex builder
would not properly build, because the travis badge link was
downloaded as an svg and latex doesn't support svgs.

Following the advice on sphinx-doc/sphinx#4720, I used
a png image, but it didn't play well with the latex formatting.
So now the travis build badge is only included in html (web)
documentation.