Ensure valid xUnits for depletedMaterial; Docs
Bug that caused #97 was a bad default choice if no
argument for timePoints was given. Now, the only options
for xUnits are days and burnup, and the correct vector
of all points is chosen, depending on the value of xUnits.
This vector is what timePoints is compared against if
actual values are pased.

Upadted the documentation, both jupyter notebook and rst.
Jupyter notebook now creates the plot that spawned this
bug and will be used in the automated testing procedure.