Cannot create a universe with negative burnup fix
Keys presented in the BranchContainer.universes dictionary
now are always positive, but the attribute hasDays reflects
the units of the burnup. BranchContainer now checks that
all universes added from file are of the same "units", i.e.
nothing switches in the middle of reading.

HomogUniv objects will not accept negative values of
bu, step, and days either.