Updated doc theme; Added changelog;
Made some changes to the basic alabaster sphinx theme to
really improve the look and feel. Consistent side-bar
showing where the whole TOC. Travis badge built-in.
Persistent GH link.

Minor tweaks to some links in the detector example.

Added a changelog that can be used to build release
notes.

Removed the pdf manual, as we are transitioning to
documentation via readthedocs