Documentation via ReadTheDocs (#102)
PDF manual removed in favor of hosting documentation at
readthedocs. Updated some features of the html theme and
external link extension in docs/conf.py to improve the look and
feel of the documentation.

A change log is included in docs/welcome/changelog.rst that should
be used to post links to major changes and bug fixes.