Default to warning logging level; Quieter operation (#106)
Default logging level is now WARNING applied to the
logger in messages.py. This is really the default, not
what is listed in settings.rc['verbosity'] at launch.

Base reader now makes all logging calls regarding
when the files are being read and completed.
This provides a single place for changes to be made,
i.e. changing the log level to a debug call, and all
changes will be inherited by all readers during the
public read method.

Quieter logging calls in parsers/__init__.py