[Docs] Default settings and variable groups (#111)
Added some scripts and documentation that explain the default settings 
and variable groups.
Scripts convert the default settings and variable groups into something 
that can be well understood by the sphinx documentation builder for
making nice doc pages. Settings and groups are also hyperlinked so 
that they can be referenced from other parts of the documentation.