No more single-entry tuples as branch keys (#114)
Fixes #113

If a branch is found to have only a single name, then the
name of the branch, and key in the branches dictionary, is taken
to be the first entry. The previous implementation would place
these branches under a single entry tuple, which required
extraneous keystrokes for retrieval. Now, the string name of the
branch is the key.

BranchReader still supports accessing branches with single entry
tuples, but this feature will be removed in the next release. A 
future warning is implemented if such a branch is accessed using
this old manner, i.e. [('nom'), ] vs. ['nom']

Also fixed a minor bug on the homogenized universe container, where
the b1Kinf data would be stored in the infinite medium dictionaries.