Extra warnings and input checking for samplers (#121)
If a globbed argument was passed into a Sampler
constructor, then no files would be loaded onto
the object. This would not cause any problems during
the read process, until now.

Two changes included in this commit. One, during
the extendFiles function, if a file glob
does not return any files, a warning is raised
indicating as much. Second, following this function call,
if no files are to be stored on the sampler and to be