Update variables for 2.1.29/2.1.30 in variables.yaml
Compared variables present in results files for the two
versions and determined the following differences.
Below is essentially `diff 2.1.29 2.1.30`, meaning
variables preceeded with > are unique to 2.1.30,
and variables with < are unique to 2.1.29.

> BALA_LOSS_NEUTRON_ERR
< BETA_DECAY_SOURCE
> DOUBLE_INDEXING
> ELECTRON_DECAY_SOURCE
> SIX_FF_EPSILON
> SIX_FF_ETA
> SIX_FF_F
> SIX_FF_KEFF
> SIX_FF_KINF
> SIX_FF_LF
> SIX_FF_LT
> SIX_FF_P

Apply these changes to the variables.yaml file and
removed a lot of un-needed yaml anchors, `&`.

Update the docs/variableGroups.rst file accordingly

Rename the `xs-yields` group to be `poisons` to better
reflect what is presented in the resutls file.
`xs-yields` is still supported through anchoring and
aliasing, but will be deprecated by 0.5.0