Add setting xs.reshapeScatter; HomogUniv reshapes scatter matrices
HomogUniv containers now reshape incoming scattering matrices
depending upon the value of settings xs.reshapeScatter.
Number of groups is inferred from the first value added
to the object, and all incoming values are compared
against this group structure. If the number of groups for
the incoming value does not match the current group
structure, a warning messages is raised, but the value
is still stored.