Fewer warnings printed for HomogUniv.addData; Convert incoming to array
The HomogUniv container will now attempt to convert non-numpy
array values to arrays. This is done because the method for
determining number of groups relies upon the shape command.

Fewer warnings are printed now for variables having different
group structures. If an array comes in that should be converted
to a matrix, it's shape will be different than the current group
structure would indicate.