Reshape scatter mats (#141)
* Add setting xs.reshapeScatter; HomogUniv reshapes scatter matrices

HomogUniv containers now reshape incoming scattering matrices
depending upon the value of settings xs.reshapeScatter.
Number of groups is inferred from the first value added
to the object, and all incoming values are compared
against this group structure. If the number of groups for
the incoming value does not match the current group
structure, a warning messages is raised, but the value
is still stored.

* BranchingReader no longer stores empty arrays

If no values are presented in a line, such as
    INF_REM 0
then the branching reader will not attempt to
store this variable. A debug messages is printed

* Rework test_containers to work with numpy arrays
* Fewer warnings printed for HomogUniv.addData; Convert incoming to array

The HomogUniv container will now attempt to convert non-numpy
array values to arrays. This is done because the method for
determining number of groups relies upon the shape command.

Fewer warnings are printed now for variables having different
group structures. If an array comes in that should be converted
to a matrix, it's shape will be different than the current group
structure would indicate.

* Direct mapping from rc __setitem__ to setValue
* Add xs.reshapeScatter to defaultSettings.rst; Update changelog