Add expected data for sensitivity tester
Using python prettyprint module (pprint), exported the
data that has been compared against MATLAB to two dictionaries.
One for sensitivities and another for the energy integrated
quantities.
These are used as the expected data for the tester.
Also add the expected parameters such as number of perturbations
and isotopes, but more data needs to be included still.