Add implicit truth methods for HomogUniv objects
HomogUniv method hasData returns True is some data is stored
on the instance. The method iterates over inf and b1 dictionaries,
expected and uncertainty, as well as metadata, and returns True
at the first instance of a non-empty dictionary.

This method is also shared with __bool__ and __nonzero__ so that
the universes can be compared implicitly, i.e. `if univ:`, without
having to call `hasData`.