[ENH] Add plot method for sensitivity reader
Plot method allows plotting of quantities stored
in the sensitivity dictionary with respect to some
or all perturbations.

Plot method accepts arguments to filter perturbations,
individual or a list of materials, zais, or perturbations can
be passed as optional keywords. If no additional keywords
are given, then all permuations of perturbations are plotted
and the plot is very busy.

Also added a label formatter, like that for depleted materials,
that can be used to automatically format the plots.
Formatter accepts special keys for materials, zai,
perturbations, and the sensitivity being plotted.