[BUG] Capture total zai directly as total
Previously, if the total zai perturbation was included as
    SENS_ZAI_LIST = [
    0    % total
    ...

This would convert the corresponding key in the self.zais to be
"0    % tota" due to the extraction process.
The extraction process is now smarter, and takes the comment
into consideration. Now, all isotopes are stored as their
original integers, while this total is stored as a "total"