ENH/DOC Add and document capability to place legends outside
Function plot.placeLegend can be used to place a legend
outside a plot, either to the right of or above. Also
supports an integer number of columns as well. This function
is added to docs/api/plot.rst as well.

Update notebook with this functionality, and include some
new figures demonstrating the feature.