API Add BaseObject for comparing; All objects inherit
This class really provides the foundation for other
readers and containers for compare against one another.
Provides a public compare method that accepts
another object, relative and absolute tolerances,
sigma for number of confidence intervals, and a
strictness parameter. This parameter determines if
warnings or exceptions are raised if one object is missing
data from another.