DOC Use inheritance to show all methods for public objects
Previous documentation showed a lot of our mixin and
subclasses, which the end user doesn't need to interact with.
By instructing sphinx to show inherited methods for classes,
we don't have to display these potentially confusing
classes.

Apply these changes to a few examples and docstrings, mainly
for depletion and detector related objects, as they
used the most subclassing