DOC Add a broad overview of the files supported and readers
File `docs/overview.rst` gives an overview of all the
files that this package currently supports, links to
the SERPENT Wiki for input and output formats, links
to our API for relevant readers, and links to our example
notebooks.

This document is included as the second chapter in the manual,
following the main landing page.