API Replace curvilinear with cylindrical and spherical dets
Working on extending the curvilinear detectors to
accurately represent the grid structure - as serpent
allows both cylindrical and spherical detectors to be created.

Expand the detectorFactory function to handle some of the
similarities between curvilinear detectors, as they both have
an R and PHI(theta per the docs?) grid, but one has a
second angular grid structure and the other has a z grid.