TST Add test for hexagonal detector
Will have to rerun the output to have a much smaller file.
Currently have ~100+ lines dedicated to a single detector
array, which is less than ideal.

Tracked hexagonal detector file should be placed under
s/t/hexplot_det0.m, or the location of
FILE_PATH for the HexagonalDetectorTester needs to be
updated.