DOC Notify user that meshPlot will plot on square mesh
Support for non-cartesian meshs plots is an impending
feature and not included in this revision.
The user can use meshPlot to plot 2D data, but the
data will not reflect the spatial mesh for the detector.

Further testing *may* indicate that meshPlot with a
hexagonal detector may not even work. That is to be seen.