API Add warnings if non-cart mesh plots requested
While the meshPlot method can support one of the
non-cartesian indices as an argument, e.g. `xcoord` for
hexagonal detectors, the plot is still rendered as a
cartesian mesh. If the user requests a plot like this
for the hexagonal, cylindrical, or spherical detectors,
a FutureWarning is raised, indicating that we're working on
implementing a better plot.