DEV Add getCommonKeys to utils module
getCommonKeys is a function that returns keys/items
found in two iterables. Supports passing descriptions for
the two sources used for reporting in the event of errors,
and a alternative function for collecting the error. The
later feature is mainly used for testing.

Added two tests in test_utils.py for verifying this
function.