[API] Convert results.metadata items to ints/floats
Some entries in the metadata dictionary for a results
reader have been converted to ints or floats, depending
upon their contents. Many of these items reflect if
certain SERPENT settings have been turned on or off,
or set to a specific mode. Examples include, but are not
limited to, pop, cycles, debug, b1BurnupCorrection,
mpiTasks, ompThreads, seed, and more.