DEV directCompare function; Modify compare doc
directCompare function is a general function that
can be used to compare strings, ints, floats, or
arrays. This function interfaces with the various notification
functions from messages to alert the end user if quantities are
identical, close but not too different, outside tolerances,
or of different types. Boolean values are returned from the method
indicating the success or failure of the comparison.

For types that are not mentioned above, an error message is
raised indicating as much. The direct comparison between
lists or sets was not as straight forward and did not get
factored into this decision. It is up to the developers
to make sure all of our bases are covered when using this
method, or not use it when not appropriate