WIP Compare results dictionaries - getGoodShapes
Working on utils function that returns the keys
for objects of identical shapes (or types if not array),
that will be used to read between results dictionaries.
This function also stores keys that are missing from
one dictionary, keys that do not lead to matching types,
and keys that lead to objects with bad shapes. These
will be used for error reporting