[DEV] Extend splitValsUncs to n-dimensional arrays
`utils.splitValsUncs` now works for higher order arrays by
splitting along the last axis. This translates to splitting
at columns of a 2D matrix. Functionality is retained for
passing in vectors or strings.

Test added to test this feature by splitting a 2D matrix
along columns.

Possible Sticking Points
------------------------

Changed how the function handles non-array objects.
If anything other than a string is passed to the function,
it is converted to array not with `utils.str2Vec` but
by `numpy.array`. This is described in the docstring.