DEV Reduce overlapping confidence intervals to debug
Now, running a comparison with verbosity > info will only
log the following cases:

info: Values that should be identical but are under lower
warning: Values that should be identical but are under upper
error: Values that should be identical but are above upper
error: Values that do not have overlapping confidence intervals

Running a comparison with verbosity at debug will log messages
for identical and values with overlapping confidence intervals.
This can be quite a lot for a results file.

This change was made to only indicate values that may have changed
drastically between runs.