DEV Better hex radius and rotation
Orientation for patch plots must be given in radians, so a
type 3 hexagon requires rotation of pi / 2 to place the
vertical edge of the flat face perpendicular to the y-axis.

Radius for hexagons is not the same as the center to center
pitch, but is instead found by some trig.
Radius is not set to be pitch / sqrt(3).

These two changes allow for hexagons to be placed edge-to-edge
for both orientations