[API] Obtain detectors from reader with getitem
Allows detectors to be obtained directly from the
reader using the __getitem__ method. This is a
shortcut to give access to the detectors in an
easier manner. Testing out this approach for
implementation in future readers, such as
Depletion, Branching, and maybe XSPlot.

Example

    >>> r = serpentTools.read('my_detector_det0.m')
    >>> r.detectors['example'] is r['example']
    True

Update example and API documentation to highlight this
change.

Added tests for this method and for the previously un-tested
iterDets method