TST Add test-requirements file for pip
This file will be used by travis to install packages that
are not necessary for every day use of the project,
but are needed for testing. The contents of the file are

flake8>=3.1.0
pandas>=0.21.0
jupyter>=1.0
codecov==2.0.15
coverage==4.5.1

Versions for flake8 and pandas were taken to be far enough back
to contain the functionality that we require.

Flake8 will eventually be used to check for lint in the
changes pushed, or the whole project for cron builds