DEV Use coverage report at end for reporting coverage
Codecov, which we previously used, doesn't do anything
until we setup a codecov account for this organization.

Instead, for now, we will just print the report to
standard out so we can get some look of how our
coverage changes over time.

Currently at 74%, which is pretty good, but we've got some
room to grow.