CI All test scripts can be run off travis as well
The main script, tools/ci_test.sh, sets a bunch of shell
variables including
    - PY_RUNNER: either "coverage run" or native "python"
      what actual runs our python scripts
    - FLAKE_DIFF: 0 if this is a not a pull request and we should
      check the whole project for lint. 1 if this is a PR.
      We use Travis environment variables TRAVIS_PULL_REQUEST to
      check if this is PR and, if so, find the intended merge branch
      with TRAVIS_BRANCH.
    - TARGET: Intended merge branch from Travis.

Script tools/checkLint.sh looks for these last two variables and
determines what level of linting to do. If FLAKE_DIFF is zero, then
run against the whole project. If FLAKE_DIFF is 1, then run
just against the diff of the current branch vs. the target branch.

If FLAKE_DIFF is not set, then we do some checking.
If the current branch is develop or master, then there isn't really
something to diff against, so we check the whole project.
Otherwise, assume this is a feature/bug-fix branch and
check against develop.

We could probably improve this default TARGET branch in the future
by doing some more interesting git stuff, like looking for upstream/
common ancestor branches. But we don't have a need for that yet.