CI Just use plain python for testNotebooks.sh
A future PR might utilize coverage run as our python
runner for the notebook tests, but not this one.

The main issue is thus: when we run coverage run setup.py test,
python utilizes the serpentTools in that directory. But, when
we run it on our example notebooks, it utilizes the serpentTools
from the pyhton path, something like
/path/to/python/site-packages/serpentTools-<version>.

Now we have two different modules that are being covered, and
coverage doesn't want to merge these nicely.

A future PR could give us a python test function, like
serpentTools.test()
and/or force the examples to use the local serpentTools
by modifying sys.path

For now, as with the previous changes, we shall not use coverage
with our jupyter notebooks.