[CI] Better CI test scripts w/ flake8 (#215)
* TST Add test-requirements file for pip

This file will be used by travis to install packages that
are not necessary for every day use of the project,
but are needed for testing. The contents of the file are

flake8>=3.1.0
pandas>=0.21.0
jupyter>=1.0
coverage==4.5.1

Versions for flake8 and pandas were taken to be far enough back
to contain the functionality that we require.

* DOC Add guides for unit testing and de-flaking

* CI All test scripts can be run off travis as well

The main script, scripts/travis/testScript.sh, sets a bunch of shell
variables including
    - PY_RUNNER: either "coverage run" or native "python"
      what actual runs our python scripts
    - FLAKE_DIFF: 0 if this is a not a pull request and we should
      check the whole project for lint. 1 if this is a PR.
      We use Travis environment variables TRAVIS_PULL_REQUEST to
      check if this is PR and, if so, find the intended merge branch
      with TRAVIS_BRANCH.
    - TARGET: Intended merge branch from Travis.

scripts/travis/checkLint.sh looks for these last two variables and
determines what level of linting to do. If FLAKE_DIFF is zero, then
run against the whole project. If FLAKE_DIFF is 1, then run
just against the diff of the current branch vs. the target branch.

If FLAKE_DIFF is not set, then we do some checking.
If the current branch is develop or master, then there isn't really
something to diff against, so we check the whole project.
Otherwise, assume this is a feature/bug-fix branch and
check against develop.

We could probably improve this default TARGET branch in the future
by doing some more interesting git stuff, like looking for upstream/
common ancestor branches. But we don't have a need for that yet.

Also added a debug option to the main testScript
that basically prints a stack trace along the way.