API Respect not setting x/y labels for sensitivity plot
Previously there was not way to not set the x/y labels for a
sensitivity plot. If the passed value did not evaluate to True,
then the default was used. This mainly occurred during
subplotting routines. Now, passing an argument that does not
evaluate to True will not trigger the default x/y label, and
result in not setting the label