DEV Expand RR postcheck w/ an inspectData method
postCheck calls a inspectData method that does basically
all of the previous post checking.

If no results data nor metadata are stored, then the
universe inspection begins. Since it is possible to not have
group constants in an output file, e.g. ``set gcu -1``, then
we have to handle that. The setting ``results.expectGcu``
can be set to False for such a case. If a RR parses through
a file and does not find results, metadata, nor universe data,
then an error is raised. If ``results.expectGcu`` is True,
then we check for data on all universes.