[API/TST] Better x/y labels for sensitivity plots; Test sensitivity plot (#229)
* API/ENH Sort perturbations, materials, zais for sensitivity

If no arguments are given to sensitivity plots, then
a sorted list of all possible options are used.
This gives structure that can be tested against and is 
reproducible across plot calls.

* API Respect not setting x/y labels for sensitivity plot

Previously there was not way to not set the x/y labels for a
sensitivity plot. If the passed value did not evaluate to True,
then the default was used. This mainly occurred during
subplotting routines. Now, passing an argument that does not
evaluate to True will not trigger the default x/y label, and
result in not setting the label

* TST Test sensitivity plot
* API Remove titleFmt argument to sensitivity plot

Was not implemented nor influenced any settings. No action
required by the user