DEV Move univ counter to private attribute - RR
The other counters were more indicative of a position in a file,
e.g. number of times this block has been encountered.
The number of universes in the counter is different, reflecting the
number of universes present in the file, irrespective of burnup.

However, this value can be zero, which caused some integer division
to break. Fixed logic for updating the result counter and the
number of obtained results in post check