DEV Check if compare objects are iterable
Change the internal check on the type of the first object
to be
if not isinstance(type0, Iterable)
rather than
if type0 in TPL_INT or whatever

This is required because some results universe data is stored
as a numpy float. This check casts all not iterables
to be single-valued vectors instead. All other iterables
are cast to arrays.