[DEV] Direct compare all arrays (#233)
* DEV Check if compare objects are iterable

Change the internal check on the type of the first object
to be
if not isinstance(type0, Iterable)
rather than
if type0 in TPL_INT or whatever

This is required because some results universe data is stored
as a numpy float. This check casts all non-iterables
to be single-valued vectors instead. All other iterables
are cast to arrays.

* DEV directCompare status codes as variables

In order to ease maintainability and readability, the
status codes for direct compare are now included as
variables in s/u/compare.py

The values have docstrings indicating their purpose, and
are directly included in the directCompare docstring.
These values should be imported, rather than using their
numeric quantities directly

* TST More explicit declaration for checkStatus
* DEV/TST Allow directCompare to work on string arrays

This is important for the depleted material, as we have vectors
of isotope names to compare quickly.