DEV Differentiate between two log bad shapes
Previously, the function messages.logBadShapes
acted upon a dictionary of key: shape pairs for
printing several bad shapes at ones. This was used
in the getKeyMatchingShapes function.

However, this feature branch also extends the directCompare
function to return a unique error code for arrays/iterables
of different shapes.

The previous function has been refactored to be
logMapOfBadShapes
to be more explicit that this function acts upon a mapping
(dictionary) of bad shapes.

A new function, called logBadShapes, now produces a
single error message to be used in logDirectCompare
to indicate that a single quantity is presented with
different shapes between two containers.