DEV Compare metadata for depletion readers
Compare method now checks the size of the vectors in
metadata. If one of the vectors does not have the same
shape on both objects, then the test immediately stops.
Otherwise, all of the subsquent material comparisons
will fail because the arrays will be of different shapes.
This saves some time.

Provided all the metadata are of the same shape, the metadata
are then compared between the readers before comparing
all the materials.