DOC/ENH Add notice on purpose of readDataFile
Closes #244 by clearly indicating the purpose of readDataFile.
If readDataFile is used to read a file that currently exits
as directly passed, then a UserWarning is raised,
but that file is still read. The warning indicates that
read is the primary function for reading, while
readDataFile should be used only for working through examples.

This should eliminate a fair amount of confusion and frustration
early on as the examples indicate that readDataFile works by
exclusively using that function. This function will not work, however,
for reading files that do not exist in the data directory.
Therefore this warning, and notices in the documented examples,
are added to relieve this burden.