DEV Add SparseReaderMixin for determining sparse support
Helper class that can be used if a user can capitalize upon
sparse matrix support via scipy. Intended to be a mixin/helper
class that accompanies a full reader.

Added a function in s/utils hasScipy that simply returns True
if scipy is installed. This is the intended manner for checking