TST Add tests for new DepmtxReader class
Tests the new reader by analyzing the full contents of the
isotopic vectors, zai vector, and deltaT value.
Depletion matrix is tested by comparing the first and last
20 values as presented in the file. Their indicies are given as
well, and these are used to slice into the depletion matrix
to obtain non-zero values.  The test file also includes tests for
the readDepmtx function as well.