DEV Use ducktyping for SparseCSCStreamProcessor regex
If it's a string, compile it and set it to be the regular
expression. Otherwise, ensure that the object passed
has a search method, e.g. is a compiled regular expression
or something else.

This should really be used by developers, so we can be more
restrictive on the typing