ENH Capture depleted material with underscores  (#255)
* ENH Add support for materials with underscores

Closes #58 with more flexible material name parsing.

* TST Improve and update depletion tester

Add tests for reading a file with underscored materials.
Use setUpModule and tearDownModule to create and destroy a
new depletion file with the same data, but with a modified
material name.

Implement a more subclassed approach to testing the depletion
reader. This allows tests for the underscored reader to be included
with very little extra writing.

Add tests for the utilities contained in the depletion file.
These utilities break up matlab variables into material name
and variable, e.g. MAT_FUEL_ADENS -> fuel, adens
and extract the variable name from a chunk of text

* TST Silence depletion tests with LoggerMixing
* DOC Add PR to changelog; Remove notice about underscores