ENH Add more clarity into supported readers
If you attempt to pass a reader string to read, e.g.

serpentTools.read('myfilePath', 'results')

the function picks the reader class based upon the second argument.
This is a good way to enforce a reader type, regardless of file
extension.
Now, the supported flags and their resulting readers are printed in
the error message.