DOC Check for numpy during install process
Update install docs to do a check of numpy first. This can be
tricky to install, both on Windows and other. Windows requires
conda, while *nix typically require sudo powers.
The check is to fire up the terminal and see if numpy is
importable. Other dependencies are resolved during the
pip install process