DEV Classes for converting containers to matlab
Add io package to be used, currently, to write data
to other files. The first example is a Matlab writer
that builds off of scipy.io.savemat.
The depletion reader still has a saveAsMatlab method,
which now utilizes a MatlabConverter class.

The MatlabConverter looks for specific method on the container
passed in order to gather data in a dictionary. That dictionary
contains the content written to the matlab file.

The intent behind the class-based approach is that each converter
class can perform checks prior to collecting and writing data.
For the MatlabConverter, the checks are:
1. scipy >= 1.0 installed, and
2. container has the gathering method

These checks are not necessary, as some exporting may not
required additional dependencies, e.g. CSV.