CLI Add to-matlab command to easily convert files
$ python -m serpentTools to-matlab

Reads in the file using serpentTools.read, and then attempts
to use serpentTools.io.toMatlab to convert to matlab.
Supports all the options of that function, but
always passes reconvert=True. This would be an easy change, but
the average use case for this subcommand is going to be
taking an existing file, dumping it in MATLAB, and then using
existing analysis scripts.