API Add branched group constant collector
Experimental class serpentTools.xs.BranchCollector
can be used to gather group constants from a coefficient
file. The group constants are stored in multidimensional tables
with dimensions corresponding to
(universe, <perturbation vector>, burnup)

The BranchCollector stores these matrices in an xsTables
dictionary, and also in BranchUniverse objects. These store
views into the same xsTables arrays, but only for specific
universes.

This should be regarded as an experimental feature, as the
data layout and API may change. The test coverage is very
minimal in this commit, with no documentation. Documentation
and an example will be added prior to feature being merged
to develop.