DOC Instructions for installing pip development
pip supports installing from github, and this allows for
users to get the latest, cutting edge update as they happen.
Instructions for installing this way are provided, with a note
that the development branch may change the API between commits.
Users should refer to the changelog before/after updating
from the develop branch