TST Update sensitivity converter test: test toMatlab
The test object stores some data as 1D vectors, e.g.
SENS_E: (3, ), but will be read back from scipy.io.loadmat
as a 2D column or row vector, (3, 1) or (1, 3). This fails the
numpy testing, so these items have to be converted back to
1D vectors.

Scalars, like nMat, are unaffected, as numpy.testing.assert_array_equal
treats them as arrays of the same shape