API Implement BranchCollector.fromFile class method
Ease of use method for constructing a branch collector
directly from a file.
With this, the reading of the file and collecting of
branching data is contained in one method.
Method also supports passing perturbations or
unknown perturbations (e.g. None)

Add a simple test that uses just the file path
to read and collect the branched data.