DEV Empty dictionary for expand variables serpent 2.1.29
For this branch, SERPENT 2.1.29 was taken to be the
base version of SERPENT, and 2.1.29 was removed as a key
from variables.yaml
For the instance where SERPENT 2.1.29 is requested,
a KeyError was raised because 2.1.29 does not exist.
We don't allow passing any version other than 2.1.29 and 2.1.30
anyway.
By using an empty dictionary, we will always take the base
variable groups when serpent 2.1.29 is set as serpentVersion