API INCOMP Zero index burnup steps for results reader
This is a breaking change, as storage of homogenized universes
has changed. The first burnup index is now 0, not 1. This move
makes the universes zero-indexed, consistent with the python
ecosystem and the rest of this package.

Tests have been updated to reflect these changes. The initial value
of burnup and days for problems with out burnup are now zero,
rather than one.