API INCOMP Store single tally detectors with floats
When reading detectors with a single tally, the value of ``tallies``,
``errors``, and ``scores`` are stored as floats, rather than
numpy arrays.

The numpy arrays were previously of a single value, often leading
to mistaken accessing methods. Slicing doesn't work as one would
first expect. Now, these are floats and can be used with,
*hopefully*, less issues

Documentation has been updated in the changelog and the
detector objects.