DOC Use global substitutions for most readers, containers
Added rst_prolog to docs/conf.py that allow the use of global
substitutions across all files. This means we don't have to
redeclare the replace:: directives at each example, changelog revision,
and overview page. Also gives a single place to make changes
if we need to in the future. All the main readers and object are
placed here, with high level read and readDataFile as well.